# docs/api/openapi.yaml
...
openapi: 3.1.0
...
info:
    title: Footnote Backend API
    # Versioning: bump when request/response shapes, status codes, or auth requirements change; docs-only edits can skip.
    version: 1.0.0
    license:
        name: MIT + Hippocratic License v3 (HL3-CORE)
        url: docs/LICENSE_STRATEGY.md
    description: |
        Backend wire contract and integration boundary for other packages.
        Chat is the shared backend action-routing boundary for web and Discord surfaces.
        Non-2xx responses use ErrorResponse unless an operation states otherwise.
        info.version is manually maintained (see docs/api/README.md).

# Explicitly mark endpoints as unauthenticated by default.
security: []
...
servers:
    - url: https://{backendHost}
      description: Backend API base URL
      variables:
          backendHost:
              default: api.your-domain.tld

# ====================
# Tags
# ====================
...
paths:
...
    /api/chat:
        post:
            operationId: postChat
            x-codeRefs:
                - packages/backend/src/handlers/chat.ts#createChatHandler
                - packages/web/src/utils/api.ts#chatQuestion
                - packages/contracts/src/web/types.ts#PostChatRequest
                - packages/contracts/src/web/types.ts#PostChatResponse
            summary: Generate a chat response
            description: |
                Submit a transport-neutral chat request and receive the backend-selected action.
                Public web callers use Turnstile when enabled; trusted internal callers may authenticate with service headers.
                Send X-Session-Id / X-Turnstile-Token headers when needed for browser traffic.
            tags:
                - Chat
            parameters:
                - $ref: '#/components/parameters/SessionIdHeader'
                - $ref: '#/components/parameters/TurnstileTokenHeader'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ChatRequest'
            responses:
                '200':
                    $ref: '#/components/responses/ChatOk'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '405':
                    $ref: '#/components/responses/MethodNotAllowed'
                '413':
                    $ref: '#/components/responses/PayloadTooLarge'
                '429':
                    $ref: '#/components/responses/TooManyRequests'
                '500':
                    $ref: '#/components/responses/InternalServerError'
                '502':
                    $ref: '#/components/responses/BadGateway'
                '503':
                    $ref: '#/components/responses/ServiceUnavailable'
        options:
            operationId: optionsChat
            x-codeRefs:
                - packages/backend/src/handlers/chat.ts#createChatHandler
            summary: CORS preflight
            description: |
                Handle CORS preflight for the chat endpoint.
                Public endpoint; no auth or rate limits.
            tags:
                - Chat
            responses:
                '204':
                    $ref: '#/components/responses/PreflightAccepted'
...
components:
...
    securitySchemes:
        # Shared secret for trace ingestion.
        TraceToken:
            type: apiKey
            in: header
            name: x-trace-token
            description: Shared secret for trusted trace ingestion.
        TrustedServiceToken:
            type: apiKey
            in: header
            name: x-service-token
            description: Shared secret for trusted internal service callers.
        # GitHub webhook signature verification.
        GitHubWebhookSignature:
            type: apiKey
            in: header
            name: X-Hub-Signature-256
            description: GitHub webhook signature (sha256=...).

    # Data models for request/response payloads.
    schemas:
...
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
                details:
                    type: string
                retryAfter:
                    type: integer
                    description: Seconds until retry is allowed.
            required:
                - error
            additionalProperties: false
...
        ChatRequest:
            type: object
            properties:
                surface:
                    type: string
                    enum: [web, discord]
                trigger:
                    type: object
                    properties:
                        kind:
                            type: string
                            enum: [submit, direct, invoked, catchup]
                        messageId:
                            type: string
                            minLength: 1
                    required:
                        - kind
                    additionalProperties: false
                latestUserInput:
                    type: string
                    minLength: 0
                    maxLength: 3072
                    description: May be an empty string for passive catchup requests; chatPlanner will typically default these to ignore.
                conversation:
                    type: array
                    minItems: 1
                    maxItems: 64
                    items:
                        $ref: '#/components/schemas/ChatConversationMessage'
                attachments:
                    type: array
                    maxItems: 8
                    items:
                        $ref: '#/components/schemas/ChatAttachment'
                capabilities:
                    $ref: '#/components/schemas/ChatCapabilities'
                sessionId:
                    type: string
                    minLength: 1
                    maxLength: 128
                surfaceContext:
                    type: object
                    properties:
                        channelId:
                            type: string
                            minLength: 1
                        guildId:
                            type: string
                            minLength: 1
                        userId:
                            type: string
                            minLength: 1
                        requestHost:
                            type: string
                            minLength: 1
                    additionalProperties: false
            required:
                - surface
                - trigger
                - latestUserInput
                - conversation
            additionalProperties: false
        ChatResponse:
            discriminator:
                propertyName: action
                mapping:
                    message: '#/components/schemas/ChatMessageAction'
                    react: '#/components/schemas/ChatReactAction'
                    ignore: '#/components/schemas/ChatIgnoreAction'
                    image: '#/components/schemas/ChatImageAction'
            oneOf:
                - $ref: '#/components/schemas/ChatMessageAction'
                - $ref: '#/components/schemas/ChatReactAction'
                - $ref: '#/components/schemas/ChatIgnoreAction'
                - $ref: '#/components/schemas/ChatImageAction'
...
        ChatConversationMessage:
            type: object
            properties:
                role:
                    type: string
                    enum: [system, user, assistant]
                content:
                    type: string
                    minLength: 1
                authorName:
                    type: string
                    minLength: 1
                authorId:
                    type: string
                    minLength: 1
                messageId:
                    type: string
                    minLength: 1
                createdAt:
                    type: string
                    minLength: 1
                    format: date-time
            required:
                - role
                - content
            additionalProperties: false
        ChatAttachment:
            type: object
            properties:
                kind:
                    type: string
                    enum: [image]
                url:
                    type: string
                    minLength: 1
                    format: uri
                    pattern: ^https?://
                contentType:
                    type: string
                    minLength: 1
            required:
                - kind
                - url
            additionalProperties: false
        ChatCapabilities:
            type: object
            properties:
                canReact:
                    type: boolean
                canGenerateImages:
                    type: boolean
                canUseTts:
                    type: boolean
            required:
                - canReact
                - canGenerateImages
                - canUseTts
            additionalProperties: false
        ChatImageRequest:
            type: object
            properties:
                prompt:
                    type: string
                    minLength: 1
                aspectRatio:
                    type: string
                    enum: [auto, square, portrait, landscape]
                background:
                    type: string
                    minLength: 1
                quality:
                    type: string
                    enum: [low, medium, high, auto]
                style:
                    type: string
                    minLength: 1
                allowPromptAdjustment:
                    type: boolean
                followUpResponseId:
                    type: string
                    minLength: 1
                outputFormat:
                    type: string
                    enum: [png, webp, jpeg]
                outputCompression:
                    type: integer
                    minimum: 1
                    maximum: 100
            required:
                - prompt
            additionalProperties: false
        ChatMessageAction:
            type: object
            properties:
                action:
                    type: string
                    enum: [message]
                message:
                    type: string
                    minLength: 1
                modality:
                    type: string
                    enum: [text, tts]
                metadata:
                    $ref: '#/components/schemas/ResponseMetadata'
            required:
                - action
                - message
                - modality
                - metadata
            additionalProperties: true
        ChatReactAction:
            type: object
            properties:
                action:
                    type: string
                    enum: [react]
                reaction:
                    type: string
                    minLength: 1
                metadata:
                    type: 'null'
            required:
                - action
                - reaction
                - metadata
            additionalProperties: true
        ChatIgnoreAction:
            type: object
            properties:
                action:
                    type: string
                    enum: [ignore]
                metadata:
                    type: 'null'
            required:
                - action
                - metadata
            additionalProperties: true
        ChatImageAction:
            type: object
            properties:
                action:
                    type: string
                    enum: [image]
                imageRequest:
                    $ref: '#/components/schemas/ChatImageRequest'
                metadata:
                    type: 'null'
            required:
                - action
                - imageRequest
                - metadata
            additionalProperties: true

        # Traces + provenance.
        # Trace metadata payloads.
        ResponseMetadataBase:
            type: object
            properties:
                provenance:
                    type: string
                    description: Expected values are Retrieved, Inferred, or Speculative.
                riskTier:
                    type: string
                    description: Expected values are Low, Medium, or High.
                tradeoffCount:
                    type: number
                    description: Expected minimum is 0.
                chainHash:
                    type: string
                licenseContext:
                    type: string
                modelVersion:
                    type: string
                staleAfter:
                    type: string
                    format: date-time
                citations:
                    type: array
                    items:
                        $ref: '#/components/schemas/Citation'
                imageDescriptions:
                    type: array
                    items:
                        type: string
                evidenceScore:
                    type: integer
                    minimum: 1
                    maximum: 5
                    description: Optional TRACE evidence chip score.
                freshnessScore:
                    type: integer
                    minimum: 1
                    maximum: 5
                    description: Optional TRACE freshness chip score.
                temperament:
                    type: object
                    description: Optional partial TRACE temperament profile with named 1-5 integer axes.
                    properties:
                        tightness:
                            type: integer
                            description: 1 to 5, where higher means stronger concision and structural efficiency.
                            minimum: 1
                            maximum: 5
                        rationale:
                            type: integer
                            description: 1 to 5, where higher means more visible rationale and trade-off explanation.
                            minimum: 1
                            maximum: 5
                        attribution:
                            type: integer
                            description: 1 to 5, where higher means clearer sourced vs inferred boundaries.
                            minimum: 1
                            maximum: 5
                        caution:
                            type: integer
                            description: 1 to 5, where higher means stronger safeguards and overclaim restraint.
                            minimum: 1
                            maximum: 5
                        extent:
                            type: integer
                            description: 1 to 5, where higher means broader options and perspectives.
                            minimum: 1
                            maximum: 5
                    additionalProperties: false
            required:
                - provenance
                - riskTier
                - tradeoffCount
                - chainHash
                - licenseContext
                - modelVersion
                - staleAfter
                - citations
            additionalProperties: true
        ResponseMetadata:
            allOf:
                - $ref: '#/components/schemas/ResponseMetadataBase'
                - type: object
                  properties:
                      responseId:
                          type: string
                  required:
                      - responseId
...
        Citation:
            type: object
            properties:
                title:
                    type: string
                url:
                    type: string
                    format: uri
                snippet:
                    type: string
            required:
                - title
                - url
            additionalProperties: false

        # Trace ingest and expiry responses.
...
    parameters:
...
        SessionIdHeader:
            name: X-Session-Id
            in: header
            required: false
            description: Optional client-generated session id used for per-session rate limiting; when omitted, the server falls back to IP.
            schema:
                type: string
        TurnstileTokenHeader:
            name: X-Turnstile-Token
            in: header
            required: false
            description: Turnstile CAPTCHA token header when enabled.
            schema:
                type: string
...
    responses:
...
        PreflightAccepted:
            description: Preflight accepted.
        ChatOk:
            description: Chat response with metadata.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ChatResponse'
...
        BadRequest:
            description: Bad request.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        Forbidden:
            description: Request was refused.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        MethodNotAllowed:
            description: Method not allowed.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        TooManyRequests:
            # Rate limit responses include Retry-After header and retryAfter body field (seconds).
            description: Rate limit exceeded.
            headers:
                Retry-After:
                    description: Seconds until retry allowed.
                    schema:
                        type: integer
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        InternalServerError:
            description: Unexpected server error.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        PayloadTooLarge:
            description: Request payload too large.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        ServiceUnavailable:
            description: Service unavailable or not configured.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        BadGateway:
            description: Upstream service error.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...